+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * unit explorer filter enhancements: 
   - displayed value in header 
+  - commands/context menu items (configure, clear)
 
 ### Changed